requirements.txt
flask
flask_cors
flask_sqlalchemy
flask_migrate
psycopg2
python-dotenv
gunicorn